refactor: move contentful folder to /libs
