+++ src/features/home/components/modules/HomeModule.tsx
@@ -7,8 +7,8 @@ import {
   VenuesModule,
 } from 'features/home/components'
 import { RecommendationModule } from 'features/home/components/modules/RecommendationModule'
-import { BusinessPane, ExclusivityPane, OffersWithCover } from 'features/home/contentful'
-import { ProcessedModule, RecommendationPane } from 'features/home/contentful/moduleTypes'
+import { BusinessPane, ExclusivityPane, OffersWithCover } from 'libs/contentful'
+import { ProcessedModule, RecommendationPane } from 'libs/contentful/moduleTypes'
 import { isOfferModuleTypeguard, isVenuesModuleTypeguard } from 'features/home/typeguards'
 
 const UnmemoizedModule = ({
