+++ src/features/home/components/modules/RecommendationModule.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ContentTypes, DisplayParametersFields } from 'features/home/contentful'
+import { ContentTypes, DisplayParametersFields } from 'libs/contentful'
 import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
 import { analytics } from 'libs/firebase/analytics'
 import { render } from 'tests/utils'
