+++ src/features/home/components/modules/RecommendationModule.tsx
@@ -6,8 +6,8 @@ import {
   ContentTypes,
   DisplayParametersFields,
   RecommendationParametersFields,
-} from 'features/home/contentful'
-import { getPlaylistItemDimensionsFromLayout } from 'features/home/contentful/dimensions'
+} from 'libs/contentful'
+import { getPlaylistItemDimensionsFromLayout } from 'libs/contentful/dimensions'
 import { analytics } from 'libs/firebase/analytics'
 import { useGeolocation } from 'libs/geolocation'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
