+++ src/features/home/components/modules/exclusivity/ExclusivityBanner.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components/native'
 
 import { ExclusivityImage } from 'features/home/components/modules/exclusivity/ExclusivityImage'
 import { ExclusivityBannerProps } from 'features/home/components/modules/exclusivity/ExclusivityModule'
-import { ContentTypes } from 'features/home/contentful'
+import { ContentTypes } from 'libs/contentful'
 import { analytics } from 'libs/firebase/analytics'
 
 const UnmemoizedExclusivityBanner = ({
