+++ src/features/home/components/modules/exclusivity/ExclusivityOffer.native.test.tsx
@@ -5,7 +5,7 @@ import { navigate } from '__mocks__/@react-navigation/native'
 import { OfferResponse } from 'api/gen'
 import * as excluOfferAPI from 'features/home/api/useExcluOffer'
 import { ExclusivityOffer } from 'features/home/components/modules/exclusivity/ExclusivityOffer'
-import { ContentTypes } from 'features/home/contentful'
+import { ContentTypes } from 'libs/contentful'
 import { offerResponseSnap as mockOffer } from 'features/offer/fixtures/offerResponse'
 import { analytics } from 'libs/firebase/analytics'
 import { render, fireEvent } from 'tests/utils'
