+++ src/features/home/components/modules/exclusivity/ExclusivityOffer.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 import { ExclusivityImage } from 'features/home/components/modules/exclusivity/ExclusivityImage'
 import { ExclusivityBannerProps } from 'features/home/components/modules/exclusivity/ExclusivityModule'
 import { useShouldDisplayExcluOffer } from 'features/home/components/modules/exclusivity/helpers/useShouldDisplayExcluOffer'
-import { ContentTypes } from 'features/home/contentful'
+import { ContentTypes } from 'libs/contentful'
 import { analytics } from 'libs/firebase/analytics'
 import { useHandleFocus } from 'libs/hooks/useHandleFocus'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
