+++ src/features/home/components/modules/exclusivity/helpers/useShouldDisplayExcluOffer.test.ts
@@ -4,7 +4,7 @@ import { mocked } from 'ts-jest/utils'
 import { OfferResponse } from 'api/gen'
 import * as excluOfferAPI from 'features/home/api/useExcluOffer'
 import { useShouldDisplayExcluOffer } from 'features/home/components/modules/exclusivity/helpers/useShouldDisplayExcluOffer'
-import { ExclusivityPane } from 'features/home/contentful'
+import { ExclusivityPane } from 'libs/contentful'
 import { offerResponseSnap as mockOffer } from 'features/offer/fixtures/offerResponse'
 import { useMaxPrice } from 'features/search/helpers/useMaxPrice/useMaxPrice'
 import { GeoCoordinates } from 'libs/geolocation'
