+++ src/features/home/pages/Home.tsx
@@ -15,7 +15,7 @@ import { HomeHeader } from 'features/home/components/headers/HomeHeader'
 import { ThematicHomeHeader } from 'features/home/components/headers/ThematicHomeHeader'
 import { HomeBodyPlaceholder } from 'features/home/components/HomeBodyPlaceholder'
 import { HomeModule } from 'features/home/components/modules/HomeModule'
-import { ProcessedModule } from 'features/home/contentful/moduleTypes'
+import { ProcessedModule } from 'libs/contentful/moduleTypes'
 import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { analytics, isCloseToBottom } from 'libs/firebase/analytics'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
