+++ src/features/home/selectHomepageEntry.test.ts
@@ -3,7 +3,7 @@ import shuffle from 'lodash/shuffle'
 import { EligibilityType, UserProfileResponse, UserRole } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useUserHasBookings } from 'features/bookings/api/useUserHasBookings'
-import { HomepageEntry, Tag } from 'features/home/contentful'
+import { HomepageEntry, Tag } from 'libs/contentful'
 import { useSelectHomepageEntry } from 'features/home/selectHomepageEntry'
 import { Credit, getAvailableCredit } from 'features/home/services/useAvailableCredit'
 import { useRemoteConfigContext } from 'libs/firebase/remoteConfig'
