+++ src/features/offer/components/OfferBody/OfferBody.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 
 import { ReportedOffer } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
-import { getPlaylistItemDimensionsFromLayout } from 'features/home/contentful/dimensions'
+import { getPlaylistItemDimensionsFromLayout } from 'libs/contentful/dimensions'
 import { useOffer } from 'features/offer/api/useOffer'
 import { useReportedOffers } from 'features/offer/api/useReportedOffers'
 import { useSimilarOffers } from 'features/offer/api/useSimilarOffers'
