+++ src/features/venue/components/VenueOffers/VenueOffers.tsx
@@ -2,8 +2,8 @@ import React, { useMemo, useCallback } from 'react'
 import { PixelRatio } from 'react-native'
 import styled from 'styled-components/native'
 
-import { Layout } from 'features/home/contentful'
-import { getPlaylistItemDimensionsFromLayout } from 'features/home/contentful/dimensions'
+import { Layout } from 'libs/contentful'
+import { getPlaylistItemDimensionsFromLayout } from 'libs/contentful/dimensions'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { LocationType } from 'features/search/enums'
 import { SearchView } from 'features/search/types'
