+++ src/libs/firebase/analytics/analytics.ts
@@ -3,7 +3,7 @@ import { Platform } from 'react-native'
 import { IdentityCheckMethod, VenueContactModel } from 'api/gen'
 import { CookiesChoiceByCategory } from 'features/cookies/types'
 import { FavoriteSortBy } from 'features/favorites/types'
-import { ContentTypes } from 'features/home/contentful'
+import { ContentTypes } from 'libs/contentful'
 import { IdentityCheckStep } from 'features/identityCheck/types'
 import { Referrals } from 'features/navigation/RootNavigator/types'
 import { FILTER_TYPES } from 'features/search/helpers/useAppliedFilters/useAppliedFilters'
