+++ src/libs/search/parseSearchParameters.ts
@@ -1,8 +1,8 @@
 import { computeBeginningAndEndingDatetimes } from 'features/home/api/helpers/computeBeginningAndEndingDatetimes'
-import { SearchParametersFields } from 'features/home/contentful'
 import { LocationType } from 'features/search/enums'
 import { sortCategories } from 'features/search/helpers/reducer.helpers'
 import { SearchState, SearchView } from 'features/search/types'
+import { SearchParametersFields } from 'libs/contentful'
 import { GeoCoordinates } from 'libs/geolocation'
 import { getCategoriesFacetFilters } from 'libs/search/utils'
 import { SubcategoryLabelMapping } from 'libs/subcategories/types'
